Move audio in common
Moves the audio/wav package out of IO.

This will be helpful for creating a jar of just this package that can be used as a library for other audio related apps.
