+++ common/src/test/kotlin/org/wycliffeassociates/otter/common/audio/wav/CueChunkTest.kt
@@ -1,4 +1,4 @@
-package org.wycliffeassociates.otter.common.io.wav
+package org.wycliffeassociates.otter.common.audio.wav
 
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertTrue
